+++ tests/testthat/test-transformation.R
@@ -10,17 +10,17 @@ test_that("exponentiate", {
     cmp2 <- comparisons(
         mod,
         variables = "gender",
-        contrast_function = "lnratioavg",
-        transformation = exp)
+        transform_pre = "lnratioavg",
+        transform_post = exp)
     expect_equal(exp(cmp1$comparison), cmp2$comparison)
     expect_equal(exp(cmp1$conf.low), cmp2$conf.low)
     expect_equal(exp(cmp1$conf.high), cmp2$conf.high)
 
-    expect_warning(tidy(cmp2, transformation = exp))
-    expect_warning(summary(cmp2, transformation = exp))
+    expect_warning(tidy(cmp2, transform_post = exp))
+    expect_warning(summary(cmp2, transform_post = exp))
 
     tid1 <- tidy(cmp1)
-    tid2 <- tidy(cmp1, transformation = exp)
+    tid2 <- tidy(cmp1, transform_post = exp)
     expect_equal(exp(tid1$estimate), tid2$estimate)
     expect_equal(exp(tid1$conf.low), tid2$conf.low)
     expect_equal(exp(tid1$conf.high), tid2$conf.high)